+++ packages/core/src/status-indicator/ValidationStatus.ts
@@ -1,8 +1,15 @@
-export const VALIDATION_NAMED_STATUS = [
-  "error",
-  "warning",
-  "success",
-  "info",
-] as const;
+export interface ValidationStatuses {
+  error: string;
+  warning: string;
+  success: string;
+  info: string;
+}
 
-export type ValidationStatus = (typeof VALIDATION_NAMED_STATUS)[number];
+export type ValidationStatus = keyof ValidationStatuses;
+
+const ValidationStatusValues = ["error", "warning", "success", "info"] as const;
+
+/**
+ * @deprecated use ValidationStatusValues
+ */
+export const VALIDATION_NAMED_STATUS = ValidationStatusValues;