+++ site/src/components/index.ts
@@ -10,9 +10,18 @@ export {
   Panel,
   SaltProvider,
   StackLayout,
+  Table,
   Tag,
+  TBody,
+  TD,
   Text,
+  TFoot,
+  TH,
+  THead,
+  TR,
 } from "@salt-ds/core";
+
+export { Kbd } from "@salt-ds/lab";
 
 export * from "./ag-grid-doc";
 export * from "./app-header";
@@ -35,18 +44,3 @@ export * from "./keyboard-controls";
 export * from "./mdx";
 export * from "./patterns";
 export * from "./quick-links";
-
-/**
- * Salt -> Markdown/MDXProvider mappings
- */
-import {
-  Table as table,
-  TBody as tbody,
-  TD as td,
-  TFoot as tfoot,
-  TH as th,
-  THead as thead,
-  TR as tr,
-} from "@salt-ds/lab";
-
-export { table, tbody, thead, th, td, tr, tfoot };