+++ packages/highcharts-theme/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @salt-ds/highcharts-theme
 
+## 0.2.0
+
+### Minor Changes
+
+- 972e37e: Added mobile density tokens.
+
+### Patch Changes
+
+- Updated dependencies [972e37e]
+- Updated dependencies [972e37e]
+- Updated dependencies [4507919]
+  - @salt-ds/theme@1.39.0
+  - @salt-ds/core@1.56.0
+
 ## 0.1.0
 
 ### Minor Changes